fix block indexing bug.
n_block, n_block_min and n_block_max are all definedd int he coordinate
system of the capacity blocks, not just the "atually used" blocks.
dualkv pre-allocates the context and decoded kv cache length up to the
capacity. when the context blocks are not used up to the full capacity,
when indexing the blocks from context to decoded, we need to count the
unused context blocks as well: this is called virtual-to-physical block
index conversion